+++ github_action_scripts/python_setup.sh
@@ -8,7 +8,7 @@ PYTHON=${HOME}/venv/bin/python
 ${PYTHON} -VV
 ${PYTHON} -m site
 ${PYTHON} -m ensurepip --upgrade
-${PYTHON} -m pip install -U pip setuptools
+${PYTHON} -m pip install -U pip setuptools==80.3.1
 echo dumping pre-installed packages
 ${PYTHON} -m pip freeze
 echo installing pip packages